С# XML comments

+++ Assets/Scripts/ScoreUI.cs
@@ -1,23 +1,31 @@
 ﻿using UnityEngine;
 using UnityEngine.UI;
 
+/// <summary>
+/// Score UI class to show a score in score section
+/// </summary>
 public class ScoreUI : MonoBehaviour
 {
     public Text scoreText;
-    public int score = 0;
+    public float score = 0;
     public bool countEnabled = true;
 
     void Update()
     {
         if(countEnabled)
-            score += (int)Time.deltaTime;
+            //count time in seconds as a score
+            score += Time.deltaTime;
         scoreText.text = score.ToString("0");
     }
 
+    /// <summary>
+    /// Sets a new score
+    /// </summary>
+    /// <param name="newScore">New score</param>
     public void SetScore(int newScore)
     {
         PlayerPrefs.SetInt("score", newScore);
-        score = newScore;
+        score = (float)newScore;
     }
 
     void OnDisable()
